refactor code, move logic into a test_helper specific to SYNQ-Golang since all the related functions are there, don't use helpers/test_helper anymore
