+++ synq/api2_test.go
@@ -1,97 +1,30 @@
 package synq
 
 import (
-	"io/ioutil"
-	"net/http"
 	"strings"
 	"testing"
 
+	"github.com/SYNQfm/SYNQ-Golang/test_helper"
 	"github.com/stretchr/testify/require"
 )
 
+var testAssetId string
+var testVideoIdV2 string
+
 const (
-	TEST_AUTH       = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwczovL3Rlc3QuYXV0aDAuY29tLyIsInN1YiI6ImF1dGgwfDU3MjE4MjFiM2ExYWFmYmUxNTlkZGE2NSIsImF1ZCI6InRESzZBdUF0QVc0ckFySzhOSTltMXdJRW5WQU9RcjUxIiwiZXhwIjoxNDkzNDM5NTExLCJpYXQiOjE0NjE4MTcxMTF9.29JkFxoHqCRPIH2wVbT-ZNIMBK8xXLwkjbLmyWxpquE"
-	V2_INVALID_AUTH = `{"message" : "invalid auth"}`
-	V2_VIDEO_ID     = "9e9dc8c8-f705-41db-88da-b3034894deb9"
+	TEST_AUTH = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwczovL3Rlc3QuYXV0aDAuY29tLyIsInN1YiI6ImF1dGgwfDU3MjE4MjFiM2ExYWFmYmUxNTlkZGE2NSIsImF1ZCI6InRESzZBdUF0QVc0ckFySzhOSTltMXdJRW5WQU9RcjUxIiwiZXhwIjoxNDkzNDM5NTExLCJpYXQiOjE0NjE4MTcxMTF9.29JkFxoHqCRPIH2wVbT-ZNIMBK8xXLwkjbLmyWxpquE"
 )
 
-func handleV2(w http.ResponseWriter, r *http.Request) {
-	var resp []byte
-	auth := r.Header.Get("Authorization")
-	k := validateAuth(auth)
-	if k != "" {
-		w.WriteHeader(http.StatusBadRequest)
-		resp = []byte(k)
-	} else {
-		type_ := "video"
-		if strings.Contains(r.URL.Path, "assets") {
-			type_ = "asset"
-		}
-		switch r.URL.Path {
-		case "/v2/videos/" + V2_VIDEO_ID,
-			"/v2/assets/" + ASSET_ID:
-			if r.Method == "GET" || r.Method == "PUT" {
-				if type_ == "asset" {
-					resp = loadSample("asset_uploaded")
-				} else {
-					resp = loadSample("video2")
-				}
-				w.WriteHeader(http.StatusOK)
-			} else if r.Method == "DELETE" {
-				w.WriteHeader(http.StatusNoContent)
-			} else {
-				w.WriteHeader(http.StatusNotFound)
-			}
-		case "/v2/videos/" + V2_VIDEO_ID + "/assets":
-			if r.Method != "GET" {
-				w.WriteHeader(http.StatusNotFound)
-			} else {
-				resp = loadSample("asset_list")
-				w.WriteHeader(http.StatusOK)
-			}
-		case "/v2/videos",
-			"/v2/assets":
-			if r.Method != "POST" {
-				resp = loadSample(type_ + "_list")
-			} else if r.Method == "POST" {
-				if type_ == "video" {
-					bytes, _ := ioutil.ReadAll(r.Body)
-					if strings.Contains(string(bytes), "user_data") {
-						resp = loadSample("new_video2_meta")
-					} else {
-						resp = loadSample("new_video2")
-					}
-				} else if type_ == "asset" {
-					resp = loadSample("asset_created")
-					w.WriteHeader(http.StatusCreated)
-				}
-				w.WriteHeader(http.StatusCreated)
-			} else {
-				w.WriteHeader(http.StatusNotFound)
-			}
-		default:
-			w.WriteHeader(http.StatusNotFound)
-		}
-	}
-	w.Write(resp)
-}
-
-func validateAuth(key string) string {
-	if !strings.Contains(key, "Bearer ") {
-		return V2_INVALID_AUTH
-	}
-	ret := strings.Split(key, "Bearer ")
-	k := ret[1]
-	if k == "fake" {
-		return V2_INVALID_AUTH
-	}
-	return ""
+func init() {
+	testAssetId = test_helper.ASSET_ID
+	testVideoIdV2 = test_helper.V2_VIDEO_ID
+	test_helper.SetSampleDir(sampleDir)
 }
 
 func setupTestApiV2(key string) ApiV2 {
 	api := NewV2(key)
-	SetupTestServer("v2")
-	api.Url = testServer.URL
+	url := test_helper.SetupServer("v2")
+	api.SetUrl(url)
 	return api
 }
 
@@ -115,7 +48,7 @@ func TestCreate2(t *testing.T) {
 	api.Key = TEST_AUTH
 	video, err := api.Create()
 	assert.Nil(err)
-	assert.Equal(V2_VIDEO_ID, video.Id)
+	assert.Equal(testVideoIdV2, video.Id)
 }
 
 func TestGet2(t *testing.T) {
@@ -123,9 +56,9 @@ func TestGet2(t *testing.T) {
 	api := setupTestApiV2(TEST_AUTH)
 	_, err := api.GetVideo("")
 	assert.NotNil(err)
-	video, err := api.GetVideo(V2_VIDEO_ID)
+	video, err := api.GetVideo(testVideoIdV2)
 	assert.Nil(err)
-	assert.Equal(video.Id, V2_VIDEO_ID)
+	assert.Equal(testVideoIdV2, video.Id)
 	assert.Len(video.Assets, 1)
-	assert.Equal(ASSET_ID, video.Assets[0].Id)
+	assert.Equal(testAssetId, video.Assets[0].Id)
 }
