+++ synq/api_test.go
@@ -8,77 +8,36 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"strings"
 	"testing"
 
+	"github.com/SYNQfm/SYNQ-Golang/test_helper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func handleV1(w http.ResponseWriter, r *http.Request) {
-	var resp []byte
-	if r.Method == "POST" {
-		bytes, _ := ioutil.ReadAll(r.Body)
-		//Parse response body
-		v, _ := url.ParseQuery(string(bytes))
-		testValues = append(testValues, v)
-		key := v.Get("api_key")
-		ke := validKey(key)
-		if ke != "" {
-			w.WriteHeader(http.StatusBadRequest)
-			resp = []byte(ke)
-		} else {
-			switch r.RequestURI {
-			case "/v1/video/details":
-				video_id := v.Get("video_id")
-				ke = validVideo(video_id)
-				if ke != "" {
-					w.WriteHeader(http.StatusBadRequest)
-					resp = []byte(ke)
-				} else {
-					resp, _ = ioutil.ReadFile("../sample/video.json")
-				}
-			case "/v1/video/create":
-				resp, _ = ioutil.ReadFile("../sample/new_video.json")
-			case "/v1/video/upload",
-				"/v1/video/uploader",
-				"/v1/video/update",
-				"/v1/video/query":
-				paths := strings.Split(r.RequestURI, "/")
-				action := paths[len(paths)-1]
-				resp = loadSample(action + ".json")
-			default:
-				w.WriteHeader(http.StatusBadRequest)
-				resp = []byte(HTTP_NOT_FOUND)
-			}
-		}
-	}
-	w.Write(resp)
+var invalidUuid string
+var testVideoIdV1 string
+var testVideoId2V1 string
+var testApiKeyV1 string
+var uploadKey string
+
+const (
+	sampleDir = "../sample"
+)
+
+func init() {
+	invalidUuid = test_helper.INVALID_UUID
+	testVideoIdV1 = test_helper.VIDEO_ID
+	testVideoId2V1 = test_helper.VIDEO_ID2
+	testApiKeyV1 = test_helper.API_KEY
+	uploadKey = test_helper.UPLOAD_KEY
 }
 
 func setupTestApi(key string, type_ ...string) Api {
 	api := NewV1(key)
-	SetupTestServer(type_...)
-	api.Url = testServer.URL
+	url := test_helper.SetupServer(type_...)
+	api.SetUrl(url)
 	return api
-}
-
-func validKey(key string) string {
-	if len(key) != 32 {
-		return INVALID_UUID
-	} else if key != API_KEY {
-		return API_KEY_NOT_FOUND
-	}
-	return ""
-}
-
-func validVideo(id string) string {
-	if len(id) != 32 {
-		return INVALID_UUID
-	} else if id != VIDEO_ID {
-		return VIDEO_NOT_FOUND
-	}
-	return ""
 }
 
 func TestMakeReq(t *testing.T) {
@@ -118,9 +77,10 @@ func TestHandlePost(t *testing.T) {
 	form.Set("test", "value")
 	err := api.handlePost("create", form, &video)
 	assert.Nil(err)
-	assert.Len(testReqs, 1)
-	r := testReqs[0]
-	v := testValues[0]
+	reqs, vals := test_helper.GetReqs()
+	assert.Len(reqs, 1)
+	r := reqs[0]
+	v := vals[0]
 	assert.Equal("/v1/video/create", r.RequestURI)
 	assert.Equal("value", v.Get("test"))
 	assert.Equal("fake", v.Get("api_key"))
@@ -153,7 +113,7 @@ func TestParseSynqResp(t *testing.T) {
 	e = parseSynqResp(a, &resp, nil, v)
 	assert.NotNil(e)
 	assert.Equal("could not parse : <Error>\n  <Code>PreconditionFailed</Code>\n  <Message>At least one of the pre-conditions you specified did not hold</Message>\n  <Condition>Bucket POST must be of the enclosure-type multipart/form-data</Condition>\n  <RequestId>634081169DAFE345</RequestId>\n  <HostId>80jHDkIWiVJd6ofogZSnvEfIxEUk35ULsvWPYFcH5f6VSUMPhCAevKwzLWN+Iw6gGTEvgogepSY=</HostId>\n</Error>\n", e.Error())
-	err_msg = []byte(INVALID_UUID)
+	err_msg = []byte(invalidUuid)
 	resp = http.Response{
 		StatusCode: 400,
 		Body:       ioutil.NopCloser(bytes.NewBuffer(err_msg)),
@@ -169,7 +129,7 @@ func TestParseSynqResp(t *testing.T) {
 	e = parseSynqResp(a, &resp, nil, v)
 	assert.NotNil(e)
 	assert.Equal("could not parse : <xml>", e.Error())
-	msg = loadSample("video.json")
+	msg = loadSample("video")
 	var video Video
 	resp = http.Response{
 		StatusCode: 200,
@@ -177,7 +137,7 @@ func TestParseSynqResp(t *testing.T) {
 	}
 	e = parseSynqResp(a, &resp, nil, &video)
 	assert.Nil(e)
-	assert.Equal(VIDEO_ID, video.Id)
+	assert.Equal(testVideoIdV1, video.Id)
 	assert.NotEmpty(video.Input)
 }
 
@@ -188,14 +148,14 @@ func TestCreate(t *testing.T) {
 	_, e := api.Create()
 	assert.NotNil(e)
 	assert.Equal("Invalid uuid. Example: '1c0e3ea4529011e6991554a050defa20'.", e.Error())
-	api.Key = API_KEY
+	api.SetKey(testApiKeyV1)
 	v, e := api.Create()
 	assert.Nil(e)
 	assert.Equal("created", v.State)
 	assert.NotNil(v.CreatedAt)
 	assert.NotNil(v.UpdatedAt)
-	assert.Equal(VIDEO_ID2, v.Id)
-	// create user userdata
+	assert.Equal(testVideoId2V1, v.Id)
+	// create user userdat
 	userdata := make(map[string]interface{})
 	userdata["importer"] = make(map[string]interface{})
 	import_data := make(map[string]string)
@@ -204,10 +164,11 @@ func TestCreate(t *testing.T) {
 	userdata["importer"] = import_data
 	_, e = api.Create(userdata)
 	assert.Nil(e)
-	assert.Len(testReqs, 3)
-	assert.Len(testValues, 3)
-	req := *testReqs[2]
-	val := testValues[2]
+	reqs, vals := test_helper.GetReqs()
+	assert.Len(reqs, 3)
+	assert.Len(vals, 3)
+	req := *reqs[2]
+	val := vals[2]
 	assert.Equal("/v1/video/create", req.URL.Path)
 	data := val.Get("userdata")
 	log.Println(data)
@@ -220,13 +181,13 @@ func TestCreate(t *testing.T) {
 
 func TestQuery(t *testing.T) {
 	assert := require.New(t)
-	api := setupTestApi(API_KEY)
+	api := setupTestApi(testApiKeyV1)
 	assert.NotNil(api)
 	filter := `if (video.state == "uploaded") { return video }`
 	videos, err := api.Query(filter)
 	assert.Nil(err)
 	assert.Len(videos, 3)
-	assert.Equal(VIDEO_ID, videos[0].Id)
+	assert.Equal(testVideoIdV1, videos[0].Id)
 	assert.Equal("de98a4c92152411fbac3b7027c8f2df7", videos[1].Id)
 	assert.Equal("bb98e6cea8224ea29e6bf00e36632bdf", videos[2].Id)
 	for _, video := range videos {
@@ -239,17 +200,17 @@ func TestGetVideo(t *testing.T) {
 	assert := require.New(t)
 	api := setupTestApi("fake")
 	assert.NotNil(api)
-	_, e := api.GetVideo(VIDEO_ID)
+	_, e := api.GetVideo(testVideoIdV1)
 	assert.NotNil(e)
 	assert.Equal("Invalid uuid. Example: '1c0e3ea4529011e6991554a050defa20'.", e.Error())
-	api.Key = API_KEY
+	api.SetKey(testApiKeyV1)
 	_, e = api.GetVideo("fake")
 	assert.NotNil(e)
 	assert.Equal("Invalid uuid. Example: '1c0e3ea4529011e6991554a050defa20'.", e.Error())
-	_, e = api.GetVideo(VIDEO_ID2)
+	_, e = api.GetVideo(testVideoId2V1)
 	assert.NotNil(e)
 	assert.Equal("Video not found.", e.Error())
-	video, e := api.GetVideo(VIDEO_ID)
+	video, e := api.GetVideo(testVideoIdV1)
 	assert.Nil(e)
 	assert.Equal("uploaded", video.State)
 	assert.NotEmpty(video.Input)
@@ -264,14 +225,15 @@ func TestGetVideo(t *testing.T) {
 
 func TestUpdateVideo(t *testing.T) {
 	assert := require.New(t)
-	api := setupTestApi(API_KEY)
+	api := setupTestApi(testApiKeyV1)
 	assert.NotNil(api)
 	source := "video.userdata = {};"
-	video, e := api.Update(VIDEO_ID, source)
+	video, e := api.Update(testVideoIdV1, source)
 	assert.Nil(e)
 	val := video.Userdata["user"].(string)
 	assert.Equal("data", val)
-	v := testValues[0]
+	_, vals := test_helper.GetReqs()
+	v := vals[0]
 	src := v.Get("source")
 	assert.Equal(source, src)
 }
