+++ synq/asset_test.go
@@ -5,12 +5,12 @@ import (
 	"net/http"
 	"testing"
 
+	"github.com/SYNQfm/SYNQ-Golang/test_helper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 const (
-	ASSET_ID       = "01823629-bcf2-4c34-b714-ae21e1a4647f"
 	ASSET_TYPE     = "mp4"
 	ASSET_CREATED  = "created"
 	ASSET_UPLOADED = "uploaded"
@@ -20,8 +20,8 @@ const (
 func setupTestVideoV2() VideoV2 {
 	api := setupTestApiV2(TEST_AUTH)
 	video, _ := api.Create()
-	SetupTestServer("v2")
-	video.Api.Url = testServer.URL
+	url := test_helper.SetupServer("v2")
+	video.Api.SetUrl(url)
 	return video
 }
 
@@ -35,8 +35,8 @@ func TestGetAssetList(t *testing.T) {
 	assets, err := api.GetAssetList()
 	assert.Nil(err)
 	assert.NotEmpty(assets)
-	assert.Equal(V2_VIDEO_ID, assets[0].VideoId)
-	assert.Equal(ASSET_ID, assets[0].Id)
+	assert.Equal(testVideoIdV2, assets[0].VideoId)
+	assert.Equal(testAssetId, assets[0].Id)
 	assert.Equal(ASSET_TYPE, assets[0].Type)
 	assert.Equal(ASSET_LOCATION, assets[0].Location)
 	assert.Equal(ASSET_CREATED, assets[0].State)
@@ -49,8 +49,8 @@ func TestGetVideoAssetList(t *testing.T) {
 	err := video.GetVideoAssetList()
 	assert.Nil(err)
 	assert.NotEmpty(video.Assets)
-	assert.Equal(V2_VIDEO_ID, video.Assets[0].VideoId)
-	assert.Equal(ASSET_ID, video.Assets[0].Id)
+	assert.Equal(testVideoIdV2, video.Assets[0].VideoId)
+	assert.Equal(testAssetId, video.Assets[0].Id)
 	assert.Equal(ASSET_TYPE, video.Assets[0].Type)
 	assert.Equal(ASSET_LOCATION, video.Assets[0].Location)
 	assert.Equal(ASSET_CREATED, video.Assets[0].State)
@@ -60,8 +60,8 @@ func TestGetAsset(t *testing.T) {
 	log.Println("Testing GetAsset")
 	assert := assert.New(t)
 	video := setupTestVideoV2()
-	asset, err := video.GetAsset(ASSET_ID)
-	assert.Equal(V2_VIDEO_ID, asset.VideoId)
+	asset, err := video.GetAsset(testAssetId)
+	assert.Equal(testVideoIdV2, asset.VideoId)
 	assert.Equal(ASSET_TYPE, asset.Type)
 	assert.Equal(ASSET_UPLOADED, asset.State)
 	assert.Equal(ASSET_LOCATION, asset.Location)
@@ -72,12 +72,12 @@ func TestUpdate(t *testing.T) {
 	log.Println("Testing Update")
 	assert := assert.New(t)
 	video := setupTestVideoV2()
-	asset, _ := video.GetAsset(ASSET_ID)
+	asset, _ := video.GetAsset(testAssetId)
 	assert.NotEmpty(asset)
 	asset.State = ASSET_UPLOADED
 	err := asset.Update()
 	assert.Nil(err)
-	asset, _ = video.GetAsset(ASSET_ID)
+	asset, _ = video.GetAsset(testAssetId)
 	assert.Equal(ASSET_UPLOADED, asset.State)
 }
 
@@ -85,7 +85,7 @@ func TestDelete(t *testing.T) {
 	log.Println("Testing Delete")
 	assert := assert.New(t)
 	video := setupTestVideoV2()
-	asset, _ := video.GetAsset(ASSET_ID)
+	asset, _ := video.GetAsset(testAssetId)
 	assert.NotEmpty(asset)
 	err := asset.Delete()
 	assert.Nil(err)
@@ -95,9 +95,9 @@ func TestHandleAssetReq(t *testing.T) {
 	log.Println("Testing TestHandleAssetReq")
 	assert := assert.New(t)
 	video := setupTestVideoV2()
-	asset, _ := video.GetAsset(ASSET_ID)
+	asset, _ := video.GetAsset(testAssetId)
 	ogAsset := asset
-	url := video.Api.getBaseUrl() + "/assets/" + ASSET_ID
+	url := video.Api.getBaseUrl() + "/assets/" + testAssetId
 	err := asset.handleAssetReq("GET", url, nil)
 	assert.Nil(err)
 	assert.Equal(ogAsset, asset)
