+++ synq/video2_test.go
@@ -4,6 +4,7 @@ import (
 	"log"
 	"testing"
 
+	"github.com/SYNQfm/SYNQ-Golang/test_helper"
 	"github.com/stretchr/testify/require"
 )
 
@@ -14,7 +15,7 @@ func TestCreateAsset(t *testing.T) {
 	asset, err := video.CreateAsset(ASSET_CREATED, ASSET_TYPE, ASSET_LOCATION)
 	assert.Nil(err)
 	assert.NotNil(asset.Id)
-	assert.Equal(ASSET_ID, asset.Id)
+	assert.Equal(testAssetId, asset.Id)
 }
 
 func TestCreateOrUpdateAsset(t *testing.T) {
@@ -24,11 +25,12 @@ func TestCreateOrUpdateAsset(t *testing.T) {
 	asset := Asset{State: ASSET_CREATED, Type: ASSET_TYPE, Location: ASSET_LOCATION}
 	err := video.CreateOrUpdateAsset(&asset)
 	assert.Nil(err)
-	assert.Equal(ASSET_ID, asset.Id)
+	assert.Equal(testAssetId, asset.Id)
 	asset.State = ASSET_UPLOADED
 	err = video.CreateOrUpdateAsset(&asset)
+	reqs, _ := test_helper.GetReqs()
 	assert.Nil(err)
-	assert.Len(testReqs, 2)
-	req := testReqs[1]
+	assert.Len(reqs, 2)
+	req := reqs[1]
 	assert.Equal("PUT", req.Method)
 }
